refactor: 도메인 레이어에 Customer 폴더 생성
