+++ src/main/java/org/devcourse/assignment/domain/customer/Address.java
@@ -1,4 +1,4 @@
-package org.devcourse.assignment.domain;
+package org.devcourse.assignment.domain.customer;
 
 import jakarta.persistence.Embeddable;
 import jakarta.validation.constraints.NotBlank;
